refactor(STransform): use uint8_t instead of size_t

+++ Bundles/visu/visuOgreAdaptor/src/visuOgreAdaptor/STransform.cpp
@@ -103,9 +103,9 @@ void STransform::updating()
     {
         const ::fwData::mt::ObjectReadLock lock(fwTransform);
 
-        for(size_t lt = 0; lt < 4; ++lt)
+        for(uint8_t lt = 0; lt < 4; ++lt)
         {
-            for(size_t ct = 0; ct < 4; ++ct)
+            for(uint8_t ct = 0; ct < 4; ++ct)
             {
                 m_ogreTransform[ct][lt] = static_cast< ::Ogre::Real >(fwTransform->getCoefficient(ct, lt));
             }
